UDP search traverse NAT
Change the CMD_SEARCH message, adding a reply-to-sender-port flag
to allow for replies to traverse a NAT.

The meaning of that flag is that the recipient should ignore the
replyPort field, and instead send a reply to the apparent port
which sent the request.

A forwarder should overwrite the replyPort field, and clear this
new MustReply flag.